feature: add wallet and account manage.todo test.

+++ src/main/java/com/ylli/api/wallet/model/Wallet.java
@@ -17,7 +17,7 @@ public class Wallet {
     public Integer recharge;
 
     //分润金额
-    public Integer bonus;
+    public Double bonus;
 
     public Timestamp createTime;
